Add `on.exit()` to `heatmap.bp()`
In the current github version of vcfR,
`heatmap.bp()` resets the plot parameters if the
function successfully exits, but if it errors, then
the plot parameters are not reset.

This preserves the user's current plotting
parameters and restores them when the function
exits, whether or not an error has occured.